test: use time.After in slowWorker, int64 counts, tighter reset validation
slowWorker: replace time.NewTimer+drain with time.After; the channel is
garbage-collected when unreferenced in Go 1.23+ so no drain is needed.

harness: CountUsageEvents, CountErrorEvents, CountIdempotencyKeys now
return int64 to avoid silent truncation on the COUNT(*) result.

scenarios: tighten RateLimit-Reset window from (now-60, now+120) to
(now, now+60) to match the actual 60-second sliding window.

migration: use explicit millisecond value ('5000') for lock_timeout to
remove any ambiguity about the unit suffix.

+++ gateway/test/harness/harness.go
@@ -486,7 +486,7 @@ func (ts *TestServer) CreateCustomer(t *testing.T, email, planID string) (uuid.U
 }
 
 // CountUsageEvents returns the number of usage_events rows for customerID.
-func (ts *TestServer) CountUsageEvents(t *testing.T, customerID uuid.UUID) int {
+func (ts *TestServer) CountUsageEvents(t *testing.T, customerID uuid.UUID) int64 {
 	t.Helper()
 	if ts == nil {
 		t.Fatal("harness: CountUsageEvents called on nil TestServer")
@@ -500,11 +500,11 @@ func (ts *TestServer) CountUsageEvents(t *testing.T, customerID uuid.UUID) int {
 	if err != nil {
 		t.Fatalf("harness: count usage_events: %v", err)
 	}
-	return int(n)
+	return n
 }
 
 // CountErrorEvents returns the number of error_events rows for customerID.
-func (ts *TestServer) CountErrorEvents(t *testing.T, customerID uuid.UUID) int {
+func (ts *TestServer) CountErrorEvents(t *testing.T, customerID uuid.UUID) int64 {
 	t.Helper()
 	if ts == nil {
 		t.Fatal("harness: CountErrorEvents called on nil TestServer")
@@ -518,12 +518,12 @@ func (ts *TestServer) CountErrorEvents(t *testing.T, customerID uuid.UUID) int {
 	if err != nil {
 		t.Fatalf("harness: count error_events: %v", err)
 	}
-	return int(n)
+	return n
 }
 
 // CountIdempotencyKeys returns the number of idempotency_keys rows for customerID and key.
 // Column name idempotency_key matches the schema (migrations/0007_idempotency_keys.sql).
-func (ts *TestServer) CountIdempotencyKeys(t *testing.T, customerID uuid.UUID, idempotencyKey string) int {
+func (ts *TestServer) CountIdempotencyKeys(t *testing.T, customerID uuid.UUID, idempotencyKey string) int64 {
 	t.Helper()
 	if ts == nil {
 		t.Fatal("harness: CountIdempotencyKeys called on nil TestServer")
@@ -541,7 +541,7 @@ func (ts *TestServer) CountIdempotencyKeys(t *testing.T, customerID uuid.UUID, i
 	if err != nil {
 		t.Fatalf("harness: count idempotency_keys: %v", err)
 	}
-	return int(n)
+	return n
 }
 
 // redisPinger adapts *redis.Client to server.HealthChecker.
